backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
pymongo==4.6.0

# Async Task Processing
celery==5.3.4
redis==5.0.1
flower==2.0.1

# Google AI and ADK
google-generativeai==0.3.2
google-cloud-aiplatform==1.38.1

# OCR
paddleocr==2.7.3
paddlepaddle==2.6.2
Pillow==10.1.0

# Data Validation
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-keycloak==3.7.0

# Cloud Storage
google-cloud-storage==2.14.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
tenacity==8.2.3

# PDF Processing
PyPDF2==3.0.1
pdf2image==1.16.3

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Logging & Monitoring
loguru==0.7.2
sentry-sdk==1.39.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.12.1
flake8==6.1.0
mypy==1.7.1
